Wizard config state (#56)
* Stage 1 of wizrd component and validation

* Wizard config/state values should always be up-to-date

There is another bug with having a navigable seed reg wizard using the sidebats which is that the proceed function of some stages mainly 3 and 4 usually persist some data into the seedconfig state after validation is performed but if the user navigates away using tjhe sidebar, the proceed function would never be called therefore the data to be persisted are never persisted.

This commit takes this approach at fixing the problem: Simply calling the proceed function introduces and unwanted feature of validation error for a stage navigated away from being dispalyed in the stage bavigated to because the validation error component for a stage is triggered in the stage's proceed function instead a new function called persisteData() is introduced into the view-model for the stages that require persisting data after validation. This persistData function is invoked by the proceed function after validation is complete. This function is also invoked by the detached method of wach class extending the abstract baseStage class to persist data after validation has been complete.

* Remove the uneeded call to the persistData function

* Add back stage state update after validation has been done in the proceed function

* Remove uneeded call to update stage state

* Only set stage state in validateInput function